+++ MusicInstrument/Brass/Trumpet.java
@@ -2,7 +2,7 @@ package MusicInstrument.Brass;
 final public class Trumpet extends BrassInstrument {
     private String style;
 
-    public Trumpet(String code, String produser, float price, float weight, boolean keyed, String style) {
+    public Trumpet(String code, String produser, double price, double weight, boolean keyed, String style) {
         super(code, produser, price, weight, keyed);
         this.style = style;
     }
