+++ MusicInstrument/Percussion/Cymbal.java
@@ -1,12 +1,12 @@
 package MusicInstrument.Percussion;
 final public class Cymbal extends PercussionInstrument {
-    private float pitch;
+    private double pitch;
 
-    public Cymbal(String code, String produser, float price, float weight, float pitch) {
+    public Cymbal(String code, String produser, double price, double weight, double pitch) {
         super(code, produser, price, weight);
         this.pitch = pitch;
     }
 
-    public float getPitch() { return this.pitch; }
-    public void setPitch(float pitch) { this.pitch = pitch; }
+    public double getPitch() { return this.pitch; }
+    public void setPitch(double pitch) { this.pitch = pitch; }
 }
