+++ MusicInstrument/Strings/Guitar.java
@@ -2,7 +2,7 @@ package MusicInstrument.Strings;
 final public class Guitar extends StringsInstrument {
     private String styleType;
 
-    public Guitar(String code, String produser, float price, float weight, int stringsNumber, String styleType) {
+    public Guitar(String code, String produser, double price, double weight, int stringsNumber, String styleType) {
         super(code, produser, price, weight, stringsNumber);
         this.styleType = styleType;
     }
